+++ source/torque_vector/CMakeLists.txt
@@ -3,6 +3,6 @@ create_can_node_lib(TORQUE_VECTOR "CMSIS_G474;FREERTOS_G474")
 add_firmware_component(
     NAME torque_vector
     LINKER_SCRIPT "STM32G474RETX_FLASH"
-    LIBS "can_node_TORQUE_VECTOR;CMSIS_G474;FREERTOS_G474;PHAL_G474;HEARTBEAT"
+    LIBS "can_node_TORQUE_VECTOR;CMSIS_G474;FREERTOS_G474;PHAL_G474;HEARTBEAT;UBLOX"
 )
